move citation recovery agent logic into a separate helper file
